@@ -1,6 +1,8 @@
 import DatePicker from "@ferz/datepicker";
-import { FaPersianLocale } from "@ferz/locales";
-import { DateTime, PersianCalendar, GregorianCalendar } from "ferz";
+import FaPersianLocale from "@ferz/persian-calendar/locales/fa-IR";
+import { DateTime } from "ferz";
+import PersianCalendar from "@ferz/persian-calendar";
+import GregorianCalendar from "@ferz/gregorian-calendar";
 import "@ferz/datepicker/css";
 
 const calendar = new PersianCalendar();
@@ -9,7 +11,7 @@ const testButton = document.querySelector(".test-button") as HTMLButtonElement;
 const datepicker = new DatePicker(
   document.querySelector(".basic-test") as HTMLInputElement,
   calendar,
-  new FaPersianLocale(),
+  FaPersianLocale,
   {
     startOfWeek: 6,
     direction: "rtl",
