@@ -1,7 +1,7 @@
-import type { Calendarable } from "./Calendarable";
-import { DateTimeComponents } from "./DateTimeComponents";
+import { type Calendarable } from "../../shared/src/Calendarable";
+import { DateTimeComponents } from "../../shared/src/DateTimeComponents";
 
-export class GregorianCalendar implements Calendarable {
+export default class GregorianCalendar implements Calendarable {
   public name: string = "gregorian";
   static GREGORIAN_EPOCH = 1721425.5;
 
