@@ -1,24 +1,23 @@
-import { GregorianCalendar } from "./GregorianCalendar";
-import { PersianCalendar } from "./PersianCalendar";
-import { IslamicCalendar } from "./IslamicCalendar";
-
+import type { Calendarable } from "../../shared/src/Calendarable";
+import type { DateTimeComponents } from "../../shared/src/DateTimeComponents";
 export class Calendar {
   static PERSIAN = "persian";
   static GREGORIAN = "gregorian";
   static ISLAMIC = "islamic";
+}
 
-  static fromName(name: string) {
-    switch (name) {
-      case "persian":
-        return new PersianCalendar();
-      case "gregorian":
-        return new GregorianCalendar();
-      case "islamic":
-        return new IslamicCalendar();
-      default:
-        throw new Error(
-          "Calendar not found. Please use a predefined calendar or register your own"
-        );
-    }
+export class EmptyCalendar implements Calendarable {
+  name: string = "empty";
+  toJd(date: DateTimeComponents): number {
+    throw new Error("Method not implemented.");
+  }
+  fromJd(jd: number): DateTimeComponents {
+    throw new Error("Method not implemented.");
+  }
+  isLeap(year: number): boolean {
+    throw new Error("Method not implemented.");
+  }
+  daysInMonth(year: number, month: number): number {
+    throw new Error("Method not implemented.");
   }
 }
