@@ -6,7 +6,9 @@ const r = (p: string) => resolve(__dirname, p);
 
 export const alias: AliasOptions = {
   ferz: r("./packages/ferz/src/"),
-  "@ferz/locales": r("./packages/locales/src/"),
+  "@ferz/gregorian-calendar": r("./packages/calendars-gregorian/src/"),
+  "@ferz/perian-calendar": r("./packages/calendars-persian/src/"),
+  "@ferz/islamic-calendar": r("./packages/calendars-islamic/src/"),
   "@ferz/datepicker": r("./packages/datepicker/src/"),
 };
 
